add dripper funds
