+++ src/utils/contracts.js
@@ -1,0 +1,47 @@
+import { ethers } from 'ethers'
+import addresses from 'constants/contractAddresses'
+import ogvAbi from 'constants/mainnetAbi/ogv.json'
+import veogvAbi from 'constants/mainnetAbi/veogv.json'
+import ousdAbi from 'constants/mainnetAbi/ousd.json'
+import vaultAbi from 'constants/mainnetAbi/vault.json'
+import dripperAbi from 'constants/mainnetAbi/dripper.json'
+
+export const getContract = (address, abi, provider) => {
+  try {
+    return new ethers.Contract(
+      address,
+      abi,
+      provider
+    )
+  } catch (e) {
+    console.error(
+      `Error creating contract in [getContract] with address:${address} abi:${JSON.stringify(
+        abi
+      )}`
+    )
+    throw e
+  }
+}
+
+export const setupContracts = () => {
+  const provider = new ethers.providers.StaticJsonRpcProvider(
+    process.env.NEXT_PUBLIC_ETHEREUM_RPC_PROVIDER,
+    { chainId: parseInt(process.env.NEXT_PUBLIC_ETHEREUM_RPC_CHAIN_ID) }
+  )
+
+  const ousd = getContract(addresses.mainnet.OUSDProxy, ousdAbi, provider)
+  const vault = getContract(addresses.mainnet.Vault, vaultAbi, provider)
+  const dripper = getContract(addresses.mainnet.Dripper, dripperAbi, provider)
+  const ogv = getContract(addresses.mainnet.OGV, ogvAbi, provider)
+  const veogv = getContract(addresses.mainnet.veOGV, veogvAbi, provider)
+
+  const contractsToExport = {
+    ousd,
+    vault,
+    dripper,
+    ogv,
+    veogv,
+  }
+
+  return contractsToExport
+}
